fix(core): yanked unicode content no longer unreadable
This is mainly Windows issue with the lang setting. As in previous
IceNvim configuration, yanking on Windows / WSL uses Windows clip.exe,
which would cause unicode content to be unreadable.

Now, with a bit of extra help, an alternative, uclip.exe, is used
instead. The setup can be viewed in README. When the executable is not
present, IceNvim will fall back to the default clip.exe of Windows.