feat(plugins): drop neorg support
Quote from `neorg-exec`: neorg itself just isn't ready yet.

I have been experimenting a lot with neorg, but it is just pointless to
keep on using something that is yet to be mature, that is still buggy
and which is not compatible with org mode which is what we want to use
neorg to replace in the first place.

I do like the idea of neorg, taking a leaf out of emacs org mode and
making daring improvements, but it is just not good enough for me, yet.