feat(plugins): change nvim-tree keymaps
The key mapped for vertical and horizontal split in nvim-tree is
changed. Now, "V" is used for vertical split instead of the original "v"
and "H" is used for horizontal split instead of the original "h".

This is because using "h" as the key for horizontal split conflicts with
moving the cursor leftwards. As for changing "v" to "V", this is simply
for matching the use of upper case for horizontal splits