feat(plugins): smarter <CR> for nvim-tree
Previously, pressing <CR> in nvim tree on a file will always result in
opening it in neovim, but it is not smart enough: there are many files
that are pointless to open in neovim; they ought to be opened via a
system call instead.

The current commit addresses this issue and decides whether to open the
file in neovim or via an external system application based on its
extension.