perf(plugins): lazy load bufferline
It had not been possible to lazy load bufferline as it needed to be
loaded before nvim-transparent, which must not be lazy loaded, so that
the bufferline would not appear transparent. However, by copying the
opaque Normal highlight group to a new IceNormal highlight group, and by
modifying nvim_get_hl to point to IceNormal if it is querying the Normal
highlight group, it is now possible to make bufferline opaque even when
lazy loading it after nvim-transparent.