feat(plugins): replace vim-surround plugin
The plugin is currently not actively maintained and is replaced in favor
of nvim-surround. Minor changes in keymaps might occur, such as
surrounding text object with tags.

For example, what used to be `ysiw<p>` (wrapping the word with a <p>
tag) nows needs to be done with `ysiwtp<CR>` (a propmt for the tag shows
up after pressing t and one simply needs to enter the tag and press
enter to finish the wrapping).