feat(lsp): unify insert / command mode cmp keymap
BREAKING CHANGE: as of this commit, we no longer use TAB / S-TAB for
selecting next / previous item during command line completion. Instead,
the keymap is now the same as that in insert mode, which is using C-j /
C-k for selecting next / previous item and using TAB for confirming
selection. A-. and A-, are also avaiable in command line now.