feat(lsp): simplify cmp keymap customization
BREAKING CHANGE: the old way `Ice.keymap.lsp.cmp` works changes
completely! Instead of being a function that returns a nvim-cmp
compatible keymap, it now becomes a table that maps the key combination
to a command. The command is a simple string, e.g. "prev_item", which
will be parsed by the setup of nvim-cmp.