perf(plugins): use VeryLazy for nvim-treesitter
The old code was for an old version of nvim-treesitter. While it could
solve the problem back then, it would significantly slow things up if we
open a file directly with nvim. The issue has been solved with a newer
version of nvim-treesitter and this piece of code is thus no longer
necessary.