perf(core): disable syntax
While improving performance slightly, this does give rise to the problem
when opening a file whose treesitter parser is not installed. However,
this should not be too much of a trouble. We can install a treesitter
manually anyway, which is why customization is allowed for IceNvim users
in the first place, and it is not too easy to discern whether the current
highlighting is treesitter- or syntax-based (at least for me) with
syntax enabled.