feat(lsp/plugins): replace nvim-cmp with blink-cmp
BREAKING CHANGE: completion engine is changed to blink-cmp. While most
behavior is expected to remain the same if you stick with the default,
things might be different here and there if you make further
customization yourself.

Pretty much everything that you configure for nvim-cmp will cease to
work. If you have heavily modified the defaults of nvim-cmp provided by
IceNvim, it is best for you to take a detailed look at the content of
this commit as well as the documentation of blink-cmp.

Apart from that, luasnip is no longer used. For any custom snippets, you
can either place them in `lua/custom/`. Note that only VSCode style
snippets are allowed now.