chore(lsp/plugins): add blink source when plugin is loaded
It is unwise to place sources that come with fittencode / orgmode
directly into the config of blink-cmp. That way, when either of these
two plugins is missing, blink would not be able to load either. The
better choice would be to place the addition of these sources in the
config of the plugins.