fix(lsp): start lsp again if first attempt fails
Namely, if a server does not exist for the first file we open so that
lsp does not start properly, we ought to do the same check again if we
are to open another file. Otherwise, if this second file does have its
corresponding servers, they would not be started nonetheless.